Add an S3 module.
This creates the following resources.

* An S3 bucket
* A resource to block public access
* AES256 encryption
* Bucket versioning
* A bucket policy from a variable

There is a variable to control whether a logging bucket is created. If
this is true, it creates the bucket with the same settings as above
except there is a separate variable for the bucket policy.

There is a GitHub action which runs git secrets against the repo and
sends to Slack if they're found. This is closing the door after the
horse has bolted but at least we'll know.
It also runs terraform fmt --check which will error if terraform is not
formatted correctly.