Add policy to allow lambda to be placed in VPC.
I've added in the policy which is required to place the lambda in the
VPC. The main reason is that terraform won't create the role policy
attachments before creating the lambda so if you try and pass these
permissions in, they won't be available to create the lambda. So they
need to be there in the first place.

You can try to set a depends_on on the policies variable but that can
cause circular dependencies if you use the lambda_arn anywhere else.

This is slightly helpful in that we don't have to pass the same
permissions for each lambda we need.