feat: Add default policy.
This will create the admin role for the key and allow it to administer
the key.
It will allow the root user all permissions on the key
It will add in a list of roles who will get encrypt/decrypt permissions.
It will add in a list of roles who will get grant access for AWS services