Add log group encryption and update redrive policy
There is an optional kms key variable to encrypt the log groups.

There is an existing kms arn variable for encrypting the environment
variables. It's possible that we'd want to use different keys for log
group and env var encryption so I've kept them seperate and updated the
name for the env var one.