DR2-1161 feat: Create skeleton step function.
This module takes a step function name, a definition as a json string
and a map of policy name -> policy string

This creates a step function with that name from the definition passed in and creates a
role with the provided policies attached.

It outputs the step function arn and the role arn.