DR2-1201 Module for EventBridge rule and API destination
There's a few changes here.

There's a module for the API destination. This is because most
eventbridge rules that are being sent to Slack will go through a single
destination so we'll need to create one and then attach multiple rules
to it.

There's one for the rules. I tried to make it generic but it's tricky so
it only accepts API destinations. If anyone wants other targets, they
can add them.

I've also removed the SNS notification from the SQS module as we don't
need to send cloudwatch alarm notifications via SNS any more.