DR2-1225 Add a new variable to the SQS module
The problem with running terraform in a clean environment is that we
were using the kms key output in a conditional which terraform doesn't
like.

I've added another parameter which defaults to kms which is used for the
conditionals and doesn't rely on any module creation first. This seems
to have helped a lot.

I've also updated the variable description for the KMS variable as it
was out of date