DR2-1668 Use RATE for DLQ alarms.
The issue is that if we get a message in a DLQ where there is already a
message in the DLQ, we don't get another notification.

Repeating messages for alarms that don't change state is a bit
complicated so we'll go with this as a solution now.

The period is 60 seconds so we shouldn't get massively spammed by a huge
number of messages entering the DLQ at once.