NOJIRA Support multiple source_arn values for the same principal service (#122)
This is a hack to enable backwards compatibility for users of the Lambda module. It uses the `any` type to support `string` and `list(string)` values as input. It then converts all to `list(string)` before transforming to a map of `map(object)`. To enable backward compatibility, the first item in the list provided will have a key matching the service principal, but other items will be suffixed with `_<index>`.